Add User delete function. Refactored (#314)
