+++ lib/review/models/review-data-layer.ts
@@ -1,23 +1,61 @@
 import { filterReviewWithAI, IResult } from '../helpers'
 import sql from '@/lib/db'
-import { Review, UserUpdatedReview } from '@/util/interfaces/interfaces'
-import {
-	Message,
-	ReviewResponseStatus,
-	UserUpdateReviewResponse,
-} from '../types/Responses'
-import { checkUserCode, createUserCode } from '../user-codes'
+import { Review } from '@/util/interfaces/interfaces'
+import { ReviewResponseStatus } from '../types/Responses'
+import { createUserCode } from '../user-codes'
 import dayjs from 'dayjs'
 import isBetween from 'dayjs/plugin/isBetween'
+import { getExistingReviewsForLandlord } from '../landlords'
+import {
+	checkForLandlordSpam,
+	checkReviewsForSimilarity,
+	updateRecentReviews,
+} from '../review-text-match'
 
 dayjs.extend(isBetween)
 
-/**
- * Data service layer for the reviews service of our backend.
- * Provides methods to create, retrieve, update or handle any other CRUD operations for reviews in the database.
- */
+export async function create(
+	inputReview: Review,
+): Promise<ReviewResponseStatus> {
+	try {
+		const existingReviewsForLandlord: Review[] =
+			await getExistingReviewsForLandlord(inputReview)
+		const reviewSpamDetected: boolean = checkReviewsForSimilarity(
+			existingReviewsForLandlord,
+			inputReview.review,
+		)
 
-export async function createReview(
+		const landlordSpamDetected: boolean = await checkForLandlordSpam(
+			inputReview.landlord,
+		)
+
+		// Don't post the review to the DB if we detect spam
+		if (reviewSpamDetected || landlordSpamDetected)
+			return {
+				message:
+					'This landlord is currently under spam protection please try again later',
+				success: false,
+			}
+
+		updateRecentReviews(inputReview.landlord).catch(() =>
+			console.error(
+				`Error Updating Recent Reviews for ${inputReview.landlord}`,
+			),
+		)
+		if (process.env.NEXT_PUBLIC_ENVIRONMENT == 'development')
+			return createReview(inputReview, {
+				flagged: false,
+				flagged_reason: 'DEV REVIEW',
+			}) // Hit data layer to create review
+
+		const filterResult: IResult = await filterReviewWithAI(inputReview)
+		return createReview(inputReview, filterResult) // Hit data layer to create review
+	} catch {
+		throw new Error()
+	}
+}
+
+async function createReview(
 	inputReview: Review,
 	filterResult: IResult,
 ): Promise<ReviewResponseStatus> {
@@ -65,130 +103,5 @@ export async function createReview(
 	} catch (e) {
 		console.error('Error Creating Review')
 		throw e
-	}
-}
-
-export async function updateReview(
-	id: number,
-	review: Review,
-): Promise<Review> {
-	await sql`UPDATE review
-           SET landlord = ${review.landlord
-							.substring(0, 150)
-							.toLocaleUpperCase()},
-               country_code = ${review.country_code.toLocaleUpperCase()},
-               city = ${review.city.substring(0, 150).toLocaleUpperCase()},
-               state = ${review.state.toLocaleUpperCase()},
-               zip = ${review.zip
-									.substring(0, 50)
-									.toLocaleUpperCase()
-									.replace(' ', '')},
-               review = ${review.review},
-               repair = ${review.repair},
-               health = ${review.health},
-               stability = ${review.stability},
-               privacy = ${review.privacy},
-               respect = ${review.respect},
-               flagged = ${review.flagged},
-               flagged_reason = ${review.flagged_reason},
-               admin_approved = ${review.admin_approved},
-               admin_edited   = ${review.admin_edited},
-			   rent = ${review.rent || null},
-			   moderation_reason = ${review.moderation_reason || null},
-			   moderator = ${review.moderator},
-			   delete_date = ${review.delete_date},
-			   delete_reason = ${review.delete_reason},
-			   deleted_by = ${review.deleted_by},
-			   restore_date = ${review.restore_date},
-			   restore_reason = ${review.restore_reason},
-			   restored_by = ${review.restored_by}
-           WHERE id = ${id};`
-
-	return review
-}
-
-function isWithinLastDay(lastAttempt: Date) {
-	const currDate = new Date()
-	const yesterday = dayjs(currDate).subtract(1, 'day')
-	return dayjs(lastAttempt).isBetween(yesterday, currDate)
-}
-
-export async function userUpdateReview(
-	id: number,
-	review: UserUpdatedReview,
-	userCode: string,
-): Promise<UserUpdateReviewResponse> {
-	//Get Review Code from Review ID
-	const selectedReviews = await sql<Review[]>`
-		SELECT user_code, number_user_attempts, last_user_attempt FROM review WHERE id = ${id};
-	`
-
-	// Review not found in DB
-	if (selectedReviews.length < 1) {
-		return {
-			success: false,
-			message: Message.NOT_FOUND,
-		}
-	}
-
-	const currentReview = selectedReviews[0]
-
-	// Too Many Attempts in the last day
-	if (
-		currentReview.number_user_attempts >= 3 &&
-		isWithinLastDay(currentReview.last_user_attempt)
-	) {
-		return {
-			success: false,
-			message: Message.RATE_LIMIT,
-		}
-	}
-
-	// No user code in review
-	if (!currentReview.user_code) {
-		return {
-			success: false,
-			message: Message.NO_CODE,
-		}
-	}
-
-	//Check that code matches the review
-	const isUpdateAllowed = await checkUserCode(userCode, currentReview.user_code)
-	if (!isUpdateAllowed) {
-		await sql`UPDATE review
-           SET 	last_user_attempt = ${new Date()},
-				number_user_attempts = ${currentReview.number_user_attempts || 0 + 1},
-
-           WHERE id = ${id};`
-		return {
-			success: false,
-			message: Message.INCORRECT,
-		}
-	} else {
-		//update review data if pass
-		const filterResult: IResult = await filterReviewWithAI(review)
-		await sql`UPDATE review
-           SET landlord = ${review.landlord
-							.substring(0, 150)
-							.toLocaleUpperCase()},
-               
-               review = ${review.review},
-               repair = ${review.repair},
-               health = ${review.health},
-               stability = ${review.stability},
-               privacy = ${review.privacy},
-               respect = ${review.respect},
-               flagged = ${filterResult.flagged},
-               flagged_reason = ${filterResult.flagged_reason},
-               admin_approved = false,
-               admin_edited   = false,
-			   rent = ${review.rent || null},
-			   moderation_reason = null,
-			   moderator = null,
-           WHERE id = ${id};`
-		return {
-			success: true,
-			message: Message.SUCCESS,
-		}
 	}
 }
