+++ lib/review/models/user-report-review.ts
@@ -1,0 +1,9 @@
+import sql from '@/lib/db'
+
+export async function report(id: number, reason: string): Promise<number> {
+	reason = reason.length > 250 ? `${reason.substring(0, 250)}...` : reason
+	await sql`UPDATE review SET flagged = true, flagged_reason = ${reason}
+      WHERE id = ${id} RETURNING id;`
+
+	return id
+}
