+++ pages/api/review/delete-review.tsx
@@ -1,7 +1,7 @@
 import { runMiddleware } from '@/util/cors'
 import { NextApiRequest, NextApiResponse } from 'next'
 import { getSession, withApiAuthRequired } from '@auth0/nextjs-auth0'
-import { deleteReview } from '@/lib/review/review'
+import { deleteReview } from '@/lib/review/models/admin-delete-review'
 
 interface IBody {
 	id: number
