+++ pages/api/review/edit-review.tsx
@@ -1,8 +1,8 @@
 import { runMiddleware } from '@/util/cors'
 import { NextApiRequest, NextApiResponse } from 'next'
 import { getSession, withApiAuthRequired } from '@auth0/nextjs-auth0'
-import { update } from '@/lib/review/review'
 import { Review } from '@/util/interfaces/interfaces'
+import { updateReview } from '@/lib/review/models/admin-update-review'
 
 const EditReview = async (req: NextApiRequest, res: NextApiResponse) => {
 	const session = await getSession(req, res)
@@ -14,7 +14,7 @@ const EditReview = async (req: NextApiRequest, res: NextApiResponse) => {
 	const id = body.id
 
 	if (user && user.role === 'ADMIN') {
-		const reviews = await update(id || 0, body)
+		const reviews = await updateReview(id || 0, body)
 		res.status(200).json(reviews)
 	} else {
 		res.status(401).json({ error: 'UNAUTHORIZED' })
