+++ pages/api/review/get-reviews.tsx
@@ -1,4 +1,4 @@
-import { getReviews } from '@/lib/review/review'
+import { getReviews } from '@/lib/review/models/user-get-reviews'
 import { ReviewQuery } from '@/lib/review/types/Queries'
 import rateLimitMiddleware from '@/util/rateLimit'
 import { NextApiRequest, NextApiResponse } from 'next'
